backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
requests
beautifulsoup4
python-dotenv
psycopg2-binary
openai
pymupdf
langchain>=0.1.0